+++ packages/vsce/__tests__/__e2e__/06_zowe.cics.showAllCommandsInPalette_false.test.ts
@@ -3,18 +3,17 @@ import { countCicsCommands, sleep } from "./util/globalMocks";
 import { expect } from "chai";
 import {openSettingsJsonEditor,writeShowAllCommandsInPaletteValue} from "./util/globalMocks";
 
-describe("Command Palette when the value is set to False", () => {
-  it("should NOT display 'IBM CICS for Zowe Explorer: Purge Task' in the command palette", async function () {
+describe("Command Palette when the enviroment variable in settings.json is set to False", () => {
+  it("Should not display 'IBM CICS for Zowe Explorer: Purge Task' in the command palette when the setting is False", async function () {
     const editor1 = await openSettingsJsonEditor();
 
-    console.log("Setting it to false");
     await writeShowAllCommandsInPaletteValue(editor1, false);
     await sleep(500);
     //saving
     await editor1.save();
-    //await sleep(500);
+  
     const countFalse = await countCicsCommands();
-    console.log("Commands when setting is False:", countFalse);
+    console.log("Number of commands when the enviroment variable in settings.json is set to False(Expecting lesser than 5 commands):", countFalse);
     const command = ">IBM CICS for Zowe Explorer: Purge Task";
 
     const workbench = new Workbench();
@@ -30,6 +29,7 @@ describe("Command Palette when the value is set to False", () => {
     const labels = await Promise.all(items.map(i => i.getLabel()));
 
     // Pass the test if the command is present
+    expect(countFalse).to.be.lessThan(6);
     expect(labels).to.not.include(
       "IBM CICS for Zowe Explorer: Purge Task",
       `Command "IBM CICS for Zowe Explorer: Purge Task" should NOT be present in the palette`