+++ packages/vsce/__tests__/__e2e__/07_zowe.cics.showAllCommandsInPalette_True.test.ts
@@ -3,18 +3,17 @@ import { countCicsCommands, sleep } from "./util/globalMocks";
 import { expect } from "chai";
 import {openSettingsJsonEditor,writeShowAllCommandsInPaletteValue} from "./util/globalMocks";
 
-describe("Command Palette when the value is set to True", () => {
-  it("should display 'IBM CICS for Zowe Explorer: Purge Task' in the command palette when the setting is True", async function () {
+describe("Command Palette when the enviroment variable in settings.json is set to True", () => {
+  it("Should display 'IBM CICS for Zowe Explorer: Purge Task' in the command palette when the setting is True", async function () {
     const editor1 = await openSettingsJsonEditor();
 
-    console.log("Setting it to True");
     await writeShowAllCommandsInPaletteValue(editor1, true);
     await sleep(500);
     //saving
     await editor1.save();
-    //await sleep(500);
+  
     const countTrue = await countCicsCommands();
-    console.log("Commands when setting is True:", countTrue);
+    console.log("Number of commands when the enviroment variable in settings.json is set to True(Expecting more than 5 commands):", countTrue);
     const command = ">IBM CICS for Zowe Explorer: Purge Task";
 
     const workbench = new Workbench();
@@ -30,6 +29,7 @@ describe("Command Palette when the value is set to True", () => {
     const labels = await Promise.all(items.map(i => i.getLabel()));
 
     // Pass the test if the command is present
+    expect(countTrue).to.be.greaterThan(6);
     expect(labels).to.include(
       "IBM CICS for Zowe Explorer: Purge Task",
       `Command "IBM CICS for Zowe Explorer: Purge Task" should be present in the palette`