+++ packages/vsce/CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to the "cics-extension-for-zowe" extension will be documente
 
 ## Recent Changes
 
+- Enhancement: Added the zowe.cics.showAllCommandsInPalette setting to control the visibility of CICS commands in the command palette. [#285] (https://github.com/zowe/cics-for-zowe-client/issues/285)
+
 - Enhancement: Navigate from a program to the library from which it was loaded. [#292](https://github.com/zowe/cics-for-zowe-client/issues/292)
 
 - Enhancement: Added pagination to all resource trees. [#302](https://github.com/zowe/cics-for-zowe-client/issues/302)
@@ -263,3 +265,4 @@ All notable changes to the "cics-extension-for-zowe" extension will be documente
 - Included auto-reading of CICS Plexes and Regions if not specified in your Zowe profile.
 - Inluded compatibility with Theia, allowing for web-based use-cases of this extension.
 - Prevention of CICS Profiles leaking into the other Zowe Explorer trees.
+